# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DeliveryMode": {
        "enum": [
          "local_link"
        ],
        "title": "DeliveryMode",
        "type": "string"
      },
      "DuplicateBody": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "DuplicateBody",
        "type": "object"
      },
      "EventProfileResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "settings": {
            "$ref": "#/components/schemas/ProfileSettingsResponse"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "settings",
          "is_active",
          "revision",
          "created_at",
          "updated_at",
          "deleted_at"
        ],
        "title": "EventProfileResponse",
        "type": "object"
      },
      "EventThemeBody": {
        "additionalProperties": false,
        "description": "Complete event theme: one colour per semantic token (every key in `tokens` is required).",
        "properties": {
          "palette": {
            "description": "Swatches extracted from the background image (source=extracted)",
            "items": {
              "examples": [
                "#2F6FD6"
              ],
              "pattern": "^#[0-9A-Fa-f]{6}$",
              "type": "string"
            },
            "maxItems": 8,
            "title": "Palette",
            "type": "array"
          },
          "preset": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "source": {
            "$ref": "#/components/schemas/ThemeSource",
            "default": "custom"
          },
          "tokens": {
            "description": "Every semantic token, e.g. background, primary_bg, primary_text, ...",
            "maxProperties": 35,
            "minProperties": 35,
            "patternProperties": {
              "^[a-z_]{1,40}$": {
                "examples": [
                  "#2F6FD6"
                ],
                "pattern": "^#[0-9A-Fa-f]{6}$",
                "type": "string"
              }
            },
            "title": "Tokens",
            "type": "object"
          }
        },
        "required": [
          "tokens"
        ],
        "title": "EventThemeBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
      "ProfileSettingsBody": {
        "additionalProperties": false,
        "description": "Every field an organizer can set. Unknown fields are refused.",
        "properties": {
          "allow_surprise_me": {
            "default": false,
            "description": "Offer a random 'Surprise me' choice (needs two frames)",
            "title": "Allow Surprise Me",
            "type": "boolean"
          },
          "background_asset_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Asset Id"
          },
          "countdown_seconds": {
            "default": 5,
            "description": "Seconds counted down before each photo (whole number)",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Countdown Seconds",
            "type": "integer"
          },
          "delivery_mode": {
            "$ref": "#/components/schemas/DeliveryMode",
            "default": "local_link"
          },
          "enabled_layouts": {
            "anyOf": [
              {
                "items": {
                  "pattern": "^[a-z0-9_]{1,64}$",
                  "type": "string"
                },
                "maxItems": 16,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Photo sizes (layout keys) participants may use; every valid frame of these sizes is offered. Omit when creating a profile to offer every size.",
            "title": "Enabled Layouts"
          },
          "inactivity_timeout_s": {
            "default": 120,
            "maximum": 900.0,
            "minimum": 30.0,
            "title": "Inactivity Timeout S",
            "type": "integer"
          },
          "logo_asset_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Asset Id"
          },
          "mirror": {
            "default": true,
            "title": "Mirror",
            "type": "boolean"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "retake_mode": {
            "$ref": "#/components/schemas/RetakeMode",
            "default": "per_photo"
          },
          "start_button_text": {
            "default": "Start",
            "maxLength": 40,
            "minLength": 1,
            "title": "Start Button Text",
            "type": "string"
          },
          "subtitle": {
            "default": "",
            "description": "Preparation screen text",
            "maxLength": 240,
            "title": "Subtitle",
            "type": "string"
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventThemeBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to use the default preset theme"
          },
          "title": {
            "description": "Preparation screen title",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "name",
          "title"
        ],
        "title": "ProfileSettingsBody",
        "type": "object"
      },
      "ProfileSettingsResponse": {
        "additionalProperties": false,
        "description": "Stored settings; the theme is always present.",
        "properties": {
          "allow_surprise_me": {
            "default": false,
            "description": "Offer a random 'Surprise me' choice (needs two frames)",
            "title": "Allow Surprise Me",
            "type": "boolean"
          },
          "background_asset_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Asset Id"
          },
          "countdown_seconds": {
            "default": 5,
            "description": "Seconds counted down before each photo (whole number)",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Countdown Seconds",
            "type": "integer"
          },
          "delivery_mode": {
            "$ref": "#/components/schemas/DeliveryMode",
            "default": "local_link"
          },
          "enabled_layouts": {
            "items": {
              "pattern": "^[a-z0-9_]{1,64}$",
              "type": "string"
            },
            "title": "Enabled Layouts",
            "type": "array"
          },
          "inactivity_timeout_s": {
            "default": 120,
            "maximum": 900.0,
            "minimum": 30.0,
            "title": "Inactivity Timeout S",
            "type": "integer"
          },
          "logo_asset_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Asset Id"
          },
          "mirror": {
            "default": true,
            "title": "Mirror",
            "type": "boolean"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "retake_mode": {
            "$ref": "#/components/schemas/RetakeMode",
            "default": "per_photo"
          },
          "start_button_text": {
            "default": "Start",
            "maxLength": 40,
            "minLength": 1,
            "title": "Start Button Text",
            "type": "string"
          },
          "subtitle": {
            "default": "",
            "description": "Preparation screen text",
            "maxLength": 240,
            "title": "Subtitle",
            "type": "string"
          },
          "theme": {
            "$ref": "#/components/schemas/EventThemeBody"
          },
          "title": {
            "description": "Preparation screen title",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "name",
          "title",
          "theme",
          "enabled_layouts"
        ],
        "title": "ProfileSettingsResponse",
        "type": "object"
      },
      "ProfileUpdateBody": {
        "additionalProperties": false,
        "properties": {
          "allow_surprise_me": {
            "default": false,
            "description": "Offer a random 'Surprise me' choice (needs two frames)",
            "title": "Allow Surprise Me",
            "type": "boolean"
          },
          "background_asset_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Asset Id"
          },
          "countdown_seconds": {
            "default": 5,
            "description": "Seconds counted down before each photo (whole number)",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Countdown Seconds",
            "type": "integer"
          },
          "delivery_mode": {
            "$ref": "#/components/schemas/DeliveryMode",
            "default": "local_link"
          },
          "enabled_layouts": {
            "description": "Photo sizes (layout keys) participants may use",
            "items": {
              "pattern": "^[a-z0-9_]{1,64}$",
              "type": "string"
            },
            "maxItems": 16,
            "title": "Enabled Layouts",
            "type": "array"
          },
          "inactivity_timeout_s": {
            "default": 120,
            "maximum": 900.0,
            "minimum": 30.0,
            "title": "Inactivity Timeout S",
            "type": "integer"
          },
          "logo_asset_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Asset Id"
          },
          "mirror": {
            "default": true,
            "title": "Mirror",
            "type": "boolean"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "retake_mode": {
            "$ref": "#/components/schemas/RetakeMode",
            "default": "per_photo"
          },
          "revision": {
            "description": "Revision the edit was based on (optimistic lock)",
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "start_button_text": {
            "default": "Start",
            "maxLength": 40,
            "minLength": 1,
            "title": "Start Button Text",
            "type": "string"
          },
          "subtitle": {
            "default": "",
            "description": "Preparation screen text",
            "maxLength": 240,
            "title": "Subtitle",
            "type": "string"
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventThemeBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to use the default preset theme"
          },
          "title": {
            "description": "Preparation screen title",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "name",
          "title",
          "enabled_layouts",
          "revision"
        ],
        "title": "ProfileUpdateBody",
        "type": "object"
      },
...
      "RetakeMode": {
        "enum": [
          "none",
          "per_photo",
          "all"
        ],
        "title": "RetakeMode",
        "type": "string"
      },
...
      "ThemeSource": {
        "enum": [
          "preset",
          "extracted",
          "custom"
        ],
        "title": "ThemeSource",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Photobooth kiosk API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/profiles": {
      "get": {
        "operationId": "list_profiles_api_admin_profiles_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventProfileResponse"
                  },
                  "title": "Response List Profiles Api Admin Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Profiles",
        "tags": [
          "admin-profiles"
        ]
      },
      "post": {
        "operationId": "create_profile_api_admin_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileSettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile",
        "tags": [
          "admin-profiles"
        ]
      }
    },
    "/api/admin/profiles/{profile_id}": {
      "delete": {
        "description": "Soft delete: the row and its assets stay; it can be restored.",
        "operationId": "delete_profile_api_admin_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Revision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": [
          "admin-profiles"
        ]
      },
      "get": {
        "operationId": "get_profile_api_admin_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "admin-profiles"
        ]
      },
      "put": {
        "operationId": "update_profile_api_admin_profiles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "admin-profiles"
        ]
      }
    },
    "/api/admin/profiles/{profile_id}/activate": {
      "post": {
        "operationId": "activate_profile_api_admin_profiles__profile_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Profile",
        "tags": [
          "admin-profiles"
        ]
      }
    },
    "/api/admin/profiles/{profile_id}/duplicate": {
      "post": {
        "operationId": "duplicate_profile_api_admin_profiles__profile_id__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DuplicateBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Profile",
        "tags": [
          "admin-profiles"
        ]
      }
    },
    "/api/admin/profiles/{profile_id}/restore": {
      "post": {
        "operationId": "restore_profile_api_admin_profiles__profile_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Profile",
        "tags": [
          "admin-profiles"
        ]
      }
    },
...
  }
}